worker/requirements.txt
amqp==5.3.1
anyio==4.11.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
ecdsa==0.19.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
pyasn1==0.6.1
pycparser==2.23
python-dateutil==2.9.0.post0
python-jose==3.5.0
redis==6.4.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
